move engine setter to base class
